LayoutTests/http/tests/resourceLoadStatistics/third-party-cookie-with-and-without-user-interaction-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
  

--------
Frame: '<!--frame1-->'
--------
Setting cookie as third party.


--------
Frame: '<!--frame2-->'
--------
Should receive both the cookie set in a first-party context and in a third-party context.
Received cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Did not receive cookie named 'partitionedThirdPartyCookie'.
Client-side document.cookie: firstPartyCookie=value,thirdPartyCookie=value

--------
Frame: '<!--frame3-->'
--------
Setting partitioned, third party cookie.


--------
Frame: '<!--frame4-->'
--------
Should only receive partitioned, third party cookie.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Received cookie named 'partitionedThirdPartyCookie'.
Client-side document.cookie: partitionedThirdPartyCookie=value

--------
Frame: '<!--frame5-->'
--------
After user interaction, should receive non-partitioned cookies set in a first-party context and in a third-party context.
Received cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Did not receive cookie named 'partitionedThirdPartyCookie'.
Client-side document.cookie: firstPartyCookie=value,thirdPartyCookie=value
